added helper function for rendering components and did some small fixes for select test
