+++ frontend/__test__/testSetup/helper.js
@@ -1,0 +1,17 @@
+import { renderRouter, screen, fireEvent } from 'expo-router/testing-library';
+
+function render(component) {
+    const MockComponent = jest.fn(() => component);
+    renderRouter(
+        {
+            index: MockComponent,
+            'directory/a': MockComponent,
+            '(group)/b': MockComponent,
+        },
+        {
+            initialUrl: '/directory/a',
+        }
+    );
+}
+
+export { render };
